+++ apps/film-storage/CHANGELOG.md
@@ -1,1 +1,3 @@
-# v
+# v0.1 - 08/01/2013
+- List of static films
+- Premium filter